Defend against a crash in dynamically-created menus
If both the values and the short names arrays are empty,
`GetValue()` returns empty string, which  `std::ranges::find()`
does not find. The distance from array end to beginning is
therefore 0, and the subsequent index into the empty `shortNames`
causes a crash.

Detect this case and return empty string.

I tripped over this because I have some dynamic menus that are
sometimes empty. I've worked around the problem in my mod, but
I thought you might want the fix.